chore: resolve rebase conflicts and modularize tests
