+++ bc_obps/registration/api/v2/__init__.py
@@ -1,5 +1,11 @@
 # ruff: noqa: F401
-from ._operations._operation_id._registration import operation, operation_representative, submission, opted_in_operation
+from ._operations._operation_id._registration import (
+    operation,
+    operation_representative,
+    submission,
+    opted_in_operation,
+    statutory_declaration,
+)
 from ._operations import operation_id
 from . import operations
 from ._user_operators._current import operator
